# backend/openapi.yaml
openapi: 3.1.0
info:
  title: MyPath API
  version: 0.1.0
  description: |
    Backend for MyPath — a Digital Media & Information Literacy (DMIL) platform
    that helps students discover, evaluate, verify and compare information about
    careers, majors, universities and scholarships.

    This document describes the API **as it currently behaves**, and is kept
    honest by contract tests in `src/tests/openapi.test.ts`: adding a route
    without documenting it fails the build.

    ## Authentication

    Session cookies. `POST /auth/register` and `POST /auth/login` set a
    `connect.sid` cookie; send it on subsequent requests. Browser clients must
    use `credentials: "include"` — the API is CORS-restricted to `FRONTEND_URL`
    and allows credentials.

    ## Content type

    Every response is JSON, including 404s and unexpected 500s.
servers:
  - url: http://localhost:5000
    description: Local development
  - url: http://backend:5000
    description: Inside the Docker Compose network
...
paths:
...
  /auth/register:
    post:
      tags: [Auth]
      summary: Create an account and start a session
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [name, email, password]
              properties:
                name: { type: string, examples: ['Sokha Chan'] }
                email: { type: string, format: email }
                password: { type: string, format: password, minLength: 8 }
      description: |
        New accounts are always created with the `student` role. A `role` field
        in the request body is ignored — honouring it would let a client
        register itself as an admin.
      responses:
        '201':
          description: Account created; session cookie set
          headers:
            Set-Cookie:
              schema: { type: string }
              description: '`connect.sid` session cookie'
          content:
            application/json:
              schema:
                type: object
                properties:
                  message: { type: string }
                  user: { $ref: '#/components/schemas/User' }
        '400':
          description: |
            Email already registered, or the request was invalid. The API
            returns the same message for both cases.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Error' }
        '403':
          description: Already logged in
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Error' }
    get:
      tags: [Auth]
      summary: Registration endpoint probe
      description: Returns a placeholder message. Present for parity with the old server-rendered app.
      security: []
      responses:
        '200':
          description: Placeholder
          content:
            application/json:
              schema:
                type: object
                properties:
                  message: { type: string }
        '403': { description: Already logged in }
  /auth/login:
    post:
      tags: [Auth]
      summary: Log in and start a session
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [email, password]
              properties:
                email: { type: string, format: email }
                password: { type: string, format: password }
      responses:
        '200':
          description: Logged in; session cookie set.
          content:
            application/json:
              schema:
                type: object
                properties:
                  message: { type: string }
                  user: { $ref: '#/components/schemas/User' }
        '400':
          description: Email or password missing
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Error' }
        '401':
          description: |
            Invalid email or password. Deliberately identical for an unknown
            email and a wrong password, so accounts cannot be enumerated.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Error' }
        '403': { description: Already logged in }
        '429': { $ref: '#/components/responses/RateLimited' }
    get:
      tags: [Auth]
      summary: Login endpoint probe
      security: []
      responses:
        '200': { description: Placeholder }
        '403': { description: Already logged in }
  /auth/me:
    get:
      tags: [Auth]
      summary: The currently authenticated user
      description: |
        Called by the frontend on every page load to restore the session.
        Not rate limited — the brute-force limiter covers only login and
        register, so normal browsing never trips it.
      responses:
        '200':
          description: Current user
          content:
            application/json:
              schema:
                type: object
                properties:
                  user: { $ref: '#/components/schemas/User' }
        '401': { $ref: '#/components/responses/Unauthorized' }
...
  /auth/google:
    get:
      tags: [Auth]
      summary: Begin Google OAuth
      security: []
      responses:
        '302': { description: Redirect to Google's consent screen }
        '503':
          description: Google OAuth is not configured on the server
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Error' }
  /auth/google/callback:
    get:
      tags: [Auth]
      summary: Google OAuth callback
      description: Redirects to `FRONTEND_URL` on success, or to `/auth/signin` on failure.
      security: []
      parameters:
        - { name: code, in: query, schema: { type: string } }
        - { name: state, in: query, schema: { type: string } }
      responses:
        '302': { description: Redirect to the frontend }
...
  /saved:
    get:
      tags: [Saved]
      summary: Everything the current student has saved
      description: |
        Newest first, across all four item types. Each entry carries a `title`,
        `subtitle` and `image` so a card can be rendered without a second
        request.
      responses:
        '200':
          description: The saved list
          content:
            application/json:
              schema:
                type: array
                items: { $ref: '#/components/schemas/SavedItem' }
        '401': { $ref: '#/components/responses/Unauthorized' }
  /saved/{type}/{id}:
    parameters:
      - name: type
        in: path
        required: true
        schema: { type: string, enum: [scholarship, major, career, university] }
      - $ref: '#/components/parameters/Id'
    post:
      tags: [Saved]
      summary: Save an item
      description: Idempotent — saving twice stores one row and still returns `201`.
      responses:
        '201':
          description: Saved
          content:
            application/json:
              schema:
                type: object
                properties:
                  message: { type: string }
        '400': { $ref: '#/components/responses/BadRequest' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '404': { $ref: '#/components/responses/NotFound' }
    delete:
      tags: [Saved]
      summary: Remove a saved item
      description: Answers `204` whether or not the item was saved.
      responses:
        '204': { description: Removed }
        '400': { $ref: '#/components/responses/BadRequest' }
        '401': { $ref: '#/components/responses/Unauthorized' }
...
components:
  securitySchemes:
    sessionCookie:
      type: apiKey
      in: cookie
      name: connect.sid
      description: Session cookie issued by `/auth/register` or `/auth/login`.
  parameters:
    Id:
      name: id
      in: path
      required: true
      schema: { type: integer, minimum: 1 }
      description: Numeric primary key.
  responses:
    BadRequest:
      description: Malformed request (for example a non-numeric `id`)
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
...
    Unauthorized:
      description: No valid session
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
    NotFound:
      description: Resource does not exist
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
    RateLimited:
      description: Too many requests (10 per 15 minutes per IP on login and register)
      headers:
        RateLimit-Limit: { schema: { type: integer } }
        RateLimit-Remaining: { schema: { type: integer } }
        RateLimit-Reset: { schema: { type: integer } }
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
  schemas:
    Error:
      type: object
      properties:
        error: { type: string, examples: ['Unauthorized: Please log in to access this page.'] }
    User:
      type: object
      description: Safe user projection — never includes the password hash.
      properties:
        id: { type: integer }
        name: { type: string }
        email: { type: string, format: email }
        role: { type: string, enum: [student, admin], default: student }
        auth_provider: { type: string, enum: [local, google] }
        google_id: { type: [string, 'null'] }
        avatar_url: { type: [string, 'null'] }
        created_at: { type: string, format: date-time }
...
    SavedItem:
      type: object
      description: |
        A polymorphic saved row, joined with enough of the target record to
        render a card. `item_id` points at a different table depending on
        `item_type`.
      properties:
        user_id: { type: integer }
        item_type: { type: string, enum: [scholarship, major, career, university] }
        item_id: { type: integer }
        saved_at: { type: string, format: date-time }
        title: { type: string }
        subtitle: { type: [string, 'null'] }
        image: { type: [string, 'null'] }
...
security:
  - sessionCookie: []